Compatibility with MS Visual Studio 2010

+++ utility.cpp
@@ -119,5 +119,5 @@ void LoadSettings(short &fontSize, short &mapVisY) {
 }
 
 float Distance(int x1, int y1, int x2, int y2) {
-	return (float)sqrt(pow(abs(x1 - x2), 2) + pow(abs(y1 - y2), 2));
+	return sqrt(pow((float)abs(x1 - x2), 2) + pow((float)abs(y1 - y2), 2));
 }
